evaluation/requirements.txt
python-Levenshtein==0.12.0

certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
requests==2.23.0
urllib3==1.25.8

psutil==5.7.0
